[FEATURE] Add abstraction for activity
