+++ gradle/dependencies.gradle
@@ -17,7 +17,6 @@ ext.Versions = [
         CORE                : '1.3.0',
         KOTLIN              : '1.3.72',
         KOTLIN_POET         : '1.6.0',
-        LIVEDATA            : '2.2.0',
         MATERIAL            : '1.1.0',
         MOSHI               : '1.9.3'
 ]
@@ -34,7 +33,6 @@ ext.Libraries = [
         CORE                : "androidx.core:core-ktx:$Versions.CORE",
         KOTLIN              : "org.jetbrains.kotlin:kotlin-stdlib:$Versions.KOTLIN",
         KOTLIN_POET         : "com.squareup:kotlinpoet:$Versions.KOTLIN_POET",
-        LIVEDATA            : "androidx.lifecycle:lifecycle-livedata-ktx:$Versions.LIVEDATA",
         MATERIAL            : "com.google.android.material:material:$Versions.MATERIAL",
         MOSHI               : "com.squareup.moshi:moshi:$Versions.MOSHI",
         MOSHI_CODEGEN       : "com.squareup.moshi:moshi-kotlin-codegen:$Versions.MOSHI"
