+++ lib/framework/build.gradle
@@ -3,7 +3,6 @@ apply from: rootProject.file('gradle/android-lib.gradle')
 dependencies {
     implementation Libraries.APPCOMPAT
     implementation Libraries.KOTLIN
-    implementation Libraries.LIVEDATA
 
     api Modules.ANNOTATION
 }
