+++ sample/src/main/kotlin/com/zulu/android/sample/MainActivity.kt
@@ -1,35 +1,24 @@
 package com.zulu.android.sample
 
-import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import androidx.annotation.IdRes
 import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
-import com.zulu.android.NavigationHandler
-import com.zulu.android.NavigationManager
+import com.zulu.android.DefaultActivity
 import com.zulu.android.sample.constant.NavigationConstant
-import kotlinx.android.synthetic.main.activity_main.*
 
-class MainActivity : AppCompatActivity(), NavigationHandler {
-    override val navigationManager = NavigationManager(this, containerId, initialScreenName)
-
-    private val layoutId: Int
+class MainActivity : DefaultActivity() {
+    override val layoutId: Int
         @LayoutRes get() = R.layout.activity_main
 
-    private val containerId: Int
+    override val containerId: Int
         @IdRes get() = R.id.container
 
-    private val initialScreenName: String
+    override val toolbarId: Int
+        @IdRes get() = R.id.toolbar
+
+    override val initialScreenName: String
         get() = NavigationConstant.FIRST
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(layoutId)
-        setSupportActionBar(toolbar)
-
-        navigationManager.onCreate(savedInstanceState)
-    }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
         // Inflate the menu; this adds items to the action bar if it is present.
@@ -45,20 +34,5 @@ class MainActivity : AppCompatActivity(), NavigationHandler {
             R.id.action_settings -> true
             else -> super.onOptionsItemSelected(item)
         }
-    }
-
-    override fun onSaveInstanceState(outState: Bundle) {
-        super.onSaveInstanceState(outState)
-        navigationManager.onSaveInstanceState(outState)
-    }
-
-    override fun onBackPressed() {
-        if (!navigationManager.onBackPressed())
-            super.onBackPressed()
-    }
-
-    override fun onSupportNavigateUp(): Boolean {
-        onBackPressed()
-        return true
     }
 }
